+++ app/src/src-SessionBear/layouts/AppLayout/index.tsx
@@ -18,7 +18,7 @@ import ThemeProvider from "lib/design-system-v2/helpers/ThemeProvider";
 import { useInitializeNewUserSessionRecordingConfig } from "features/settings/components/SessionsBookSettings/hooks";
 import DBListeners from "hooks/DbListenerInit/DBListeners";
 import { Outlet } from "react-router-dom";
-import { GlobalModals } from "./GlobalModals";
+import { GlobalModals } from "components/common/GlobalModals";
 
 const App = () => {
   useEffect(() => {
